Add: d.ts file
