+++ @types/Search.d.ts
@@ -1,0 +1,7 @@
+export declare module 'Search';
+
+type searchResults = {
+  profileImage: string;
+  rankerName: string;
+  tierImage: string | null;
+};
