+++ src/pages/community/article/CommentList.tsx
@@ -1,11 +1,12 @@
 import React, { useEffect, useState } from 'react';
 import Comment from './Comment';
+import { CommentListProps } from '../../../../@types/Article';
 import './CommentList.scss';
-import { CommentListProps } from './Article';
 
 const CommentList = (props: CommentListProps) => {
   const { commentList, setCommentList, copyCommentList, loadArticleComment } =
     props;
+
   const [currentTab, setCurrentTab] = useState<number>(0);
 
   // 최신순, 인기순 탭 기능
